# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlyRank Task & Auth API",
    "version": "4.0.0",
    "description": "Secure REST API with Supabase Auth integration, Bearer JWT protection, and PostgreSQL persistence"
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing email/password or registration error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user & return JWT tokens",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully, returns JWT access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "400": {
            "description": "Missing inputs"
          },
          "401": {
            "description": "Invalid login credentials"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Terminate user session",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": {
            "description": "Logged out successfully (No Content)"
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Read public unprotected data",
        "tags": ["Public Routes"],
        "responses": {
          "200": {
            "description": "Public greeting message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Welcome stranger! This info is public." }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private authenticated user profile data",
        "tags": ["Protected Routes"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Verified user profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing, invalid, or expired token"
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Read protected dashboard data",
        "tags": ["Protected Routes"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Protected dashboard payload"
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "tags": ["Tasks API"],
        "responses": {
          "200": {
            "description": "A list of tasks"
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "tags": ["Tasks API"],
        "responses": {
          "201": {
            "description": "Task created"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Supabase JWT access token obtained from /auth/login"
      }
    },
    "schemas": {
      "AuthCredentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "example": "testuser@example.com"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "token_type": { "type": "string", "example": "bearer" },
          "expires_in": { "type": "integer", "example": 3600 },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "usr_94820a" },
          "email": { "type": "string", "example": "testuser@example.com" },
          "role": { "type": "string", "example": "authenticated" },
          "created_at": { "type": "string", "example": "2026-08-23T02:40:00.000Z" }
        }
      }
    }
  }
}